fix: conftest cleanup now don't pick up base DjangoModel class

+++ tests/conftest.py
@@ -40,7 +40,8 @@ def clean_database(request):
     ]
 
     for model in models:
-        if model._meta.app_label != "ignore":
+
+        if model._meta.app_label != "ignore" and not model._meta.abstract:
             with connection.cursor() as cursor:
                 cursor.execute(f"DELETE FROM {model._meta.db_table}")
                 cursor.execute(f"REFRESH TABLE {model._meta.db_table}")
